Replace the git image with alpine as a base image.
The Prow images using the git image
(https://github.com/kubernetes/test-infra/tree/master/images/git) use
git and openssh, and include a couple files for known hosts. Both git
and openssh are available in alpine images, and Prow should have
in-built support for known hosts in clonerefs as of a few years ago.

We should be able to replace the git image with the alpine image the
other components use instead. (Since these are the only images using the
git image, we should also be able to clean up the git image after this).